requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
packaging==19.2
psycopg2-binary==2.8.3
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
